+++ esp32/snapclient/include/mode_switch_controller.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include "runtime_mode.h"
 #include "snapclient_config.h"
 
 class ModeSwitchController {
  public:
   void begin(app_config::OperatingMode currentMode);
+  void setRuntimeMode(RuntimeMode *runtimeMode);
   void update();
 
  private:
@@ -15,6 +17,7 @@ class ModeSwitchController {
   void requestModeToggle();
 
   app_config::OperatingMode currentMode_ = app_config::OperatingMode::Snapclient;
+  RuntimeMode *runtimeMode_ = nullptr;
   bool buttonWasReleasedAfterBoot_ = false;
   bool pressHandled_ = false;
   uint32_t pressedSinceMs_ = 0;