+++ esp32/snapclient/include/runtime_mode.h
@@ -7,4 +7,5 @@ class RuntimeMode {
   virtual bool begin() = 0;
   virtual void loop() = 0;
   virtual const char *name() const = 0;
+  virtual void prepareForRestart() {}
 };